+++ src/core/composeHandlers.js
@@ -1,0 +1,9 @@
+import callHandlers from './callHandlers'
+
+function composeHandlers(handlers = []) {
+  return event => {
+    callHandlers(event, handlers)
+  }
+}
+
+export default composeHandlers